Stack of unneeded casts or unused local variables.  Thanks eclipse!
git-svn-id: http://scarab.tigris.org/svn/scarab@8933 3c7165bb-d2ef-0310-b724-bafd9a86081f

+++ trunk/src/java/org/tigris/scarab/util/ScarabUtil.java
@@ -235,8 +235,8 @@ public class ScarabUtil
             {
                 byte toEscape = bytes[i];
                 out.append('%');
-                int low = (int) (toEscape & 0x0f);
-                int high = (int) ((toEscape & 0xf0) >> 4);
+                int low = (toEscape & 0x0f);
+                int high = ((toEscape & 0xf0) >> 4);
                 out.append(HEXADECIMAL[high]);
                 out.append(HEXADECIMAL[low]);
             }
